requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
timm==0.9.2
albumentations==1.3.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.4
pillow==9.5.0
matplotlib==3.8.0
tqdm==4.66.1
streamlit==1.26.0
segmentation-models-pytorch==0.3.3
efficientnet-pytorch==0.7.1
grad-cam==1.4.8
